// Company_tasks/Aptech-Solutions_Product_Management_Dashboard/product-dashboard/src/types/auth.types.ts
export interface User {
  id: number;
  username: string;
  email: string;
  firstName: string;
  lastName: string;
  gender: string;
  image: string;
}

export interface AuthResponse extends User {
  accessToken: string; 
  refreshToken: string;
}

export interface LoginCredentials {
  username: string;
  password: string;
  expiresInMins?: number;
}